Dejank ColorInput
